@@ -37,69 +40,69 @@
   "dependencies": {},
   "devDependencies": {
     "annogenerate": "^0.8.1",
-    "babel-cli": "^6.11.4",
-    "babel-core": "^6.13.2",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-istanbul": "^2.0.0",
+    "babel-cli": "^6.16.0",
+    "babel-core": "^6.17.0",
+    "babel-eslint": "^7.0.0",
+    "babel-loader": "^6.2.5",
+    "babel-plugin-istanbul": "^2.0.1",
     "babel-plugin-syntax-object-rest-spread": "^6.13.0",
-    "babel-plugin-transform-object-rest-spread": "^6.8.0",
+    "babel-plugin-transform-object-rest-spread": "^6.16.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.2.9",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-react": "^6.11.1",
+    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-react": "^6.16.0",
     "babel-preset-react-hmre": "^1.1.1",
-    "babel-register": "^6.11.6",
+    "babel-register": "^6.16.3",
     "catalog": "bebraw/catalog#reactabular",
     "chai": "^3.5.0",
     "classnames": "^2.2.5",
-    "clean-webpack-plugin": "^0.1.10",
+    "clean-webpack-plugin": "^0.1.13",
     "copy-webpack-plugin": "^3.0.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^3.3.1",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-plugin-import": "^1.13.0",
-    "eslint-plugin-jsx-a11y": "^2.1.0",
-    "eslint-plugin-react": "^6.1.1",
+    "css-loader": "^0.25.0",
+    "eslint": "^3.7.1",
+    "eslint-config-airbnb": "^12.0.0",
+    "eslint-plugin-import": "^1.16.0",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-react": "^6.4.0",
     "expose-loader": "^0.7.1",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
     "gh-pages": "^0.11.0",
     "git-prepush-hook": "^1.0.1",
-    "html-loader": "^0.4.3",
+    "html-loader": "^0.4.4",
     "html-webpack-plugin": "^2.22.0",
-    "karma": "^1.2.0",
+    "karma": "^1.3.0",
     "karma-chai": "^0.1.0",
     "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.1.1",
-    "karma-phantomjs-launcher": "^1.0.1",
+    "karma-mocha": "^1.2.0",
+    "karma-phantomjs-launcher": "^1.0.2",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.26",
     "karma-webpack": "^1.8.0",
-    "lerna": "2.0.0-beta.25",
-    "lodash": "^4.15.0",
-    "mocha": "^3.0.2",
+    "lerna": "2.0.0-beta.30",
+    "lodash": "^4.16.4",
+    "mocha": "^3.1.1",
     "phantomjs-polyfill": "0.0.2",
-    "phantomjs-prebuilt": "^2.1.12",
+    "phantomjs-prebuilt": "^2.1.13",
     "purecss": "^0.6.0",
     "raw-loader": "^0.5.1",
-    "react": "^15.3.0",
-    "react-addons-perf": "^15.3.0",
-    "react-addons-test-utils": "^15.3.0",
+    "react": "^15.3.2",
+    "react-addons-perf": "^15.3.2",
+    "react-addons-test-utils": "^15.3.2",
     "react-dnd": "^2.1.4",
     "react-dnd-html5-backend": "^2.1.2",
-    "react-dom": "^15.3.0",
+    "react-dom": "^15.3.2",
     "react-github-corner": "^0.3.0",
     "react-pagify": "^2.1.1",
     "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
+    "redux": "^3.6.0",
     "schema2object": "^0.4.0",
     "segmentize": "^0.4.1",
     "style-loader": "^0.13.1",
-    "stylesheet-helpers": "^0.1.2",
+    "stylesheet-helpers": "^0.2.2",
     "url-loader": "^0.5.7",
-    "uuid": "^2.0.2",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
+    "uuid": "^2.0.3",
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.16.2",
     "webpack-merge": "^0.14.1"
   },
   "pre-push": [
