@@ -1,7 +1,7 @@
 import React from 'react';
 
 const SearchColumns = ({ columns, query, onChange }) => {
-  const onQueryChange = event => {
+  const onQueryChange = (event) => {
     onChange({
       ...query,
       [event.target.name]: event.target.value
@@ -10,15 +10,15 @@ const SearchColumns = ({ columns, query, onChange }) => {
 
   return (
     <tr>
-      {columns.map((column, i) => (
-        <th key={`${i}-column-filter`} className="column-filter">
+      {columns.map(column => (
+        <th key={`${column.property}-column-filter`} className="column-filter">
           {column && column.property ?
             <input
               onChange={onQueryChange}
               className="column-filter-input"
               name={column.property}
               placeholder={column.filterPlaceholder || ''}
-              value={query[i]}
+              value={query[column.property]}
             />
           : ''}
         </th>
