@@ -1,12 +1,12 @@
 import getParents from './get-parents';
 
 // TODO: Refactor the first parameter into an object
-const filterTree = (fieldName, id) => rows => rows.filter((item, index) => {
+const filterTree = fieldName => rows => rows.filter((item, index) => {
   if (!item.parent) {
     return true;
   }
 
-  const parents = getParents({ rows, index, id });
+  const parents = getParents({ rows, index });
 
   return parents.filter(
     parent => parent[fieldName]
