@@ -2,13 +2,13 @@
 function unpackTree(rows = []) {
   let ret = [];
 
-  rows.forEach(row => {
+  rows.forEach((row) => {
     const { _pack, ...rest } = row;
 
     ret = ret.concat([rest]).concat(_pack);
   });
 
-  return ret;
+  return ret.filter(a => a);
 }
 
 export default unpackTree;
