@@ -1,7 +1,7 @@
 function resolveBodyColumns(columns) {
   let ret = [];
 
-  columns.forEach(column => {
+  columns.forEach((column) => {
     // If a column has children, skip cell specific configuration
     if (column.children && column.children.length) {
       ret = ret.concat(resolveBodyColumns(column.children));
